+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **FEAT**: Add `ShadInputOTP` and `ShadInputOTPFormField` components.
 - **BREAKING CHANGE**: The `ShadBorder` no longer has default values, if you want to see the border rendered provide a color and a width greater than 0. This fixes the merge of material borders.
+- **FIX**: `ShadTabs` border radius.
 
 ## 0.16.3